group name change bug fix:
when a group's name is changed, the committee names and the page's
group_name are updated. also, the wiki's rendered html is erased so
that it will be re-rendered on the next viewing and rendered with the
new group name. this only happens when the wiki's primary group has
had its name change because only the primary group name is inserted
automatically into links. if you manually specify a link with a group
scope, then you are out of luck. in the future, if we store a record
of backlinks, we could possibly update manually scoped links as well.