Big code cleanup of the module PageFinders and new page list headings.
This will make it easier to add page filter paths in the future.
The most significant change is that now there is a parser for the
path. We can use this parser to determine the current sort order
and the current filter. The code is easier to read, or so i hope.

Also included is a first attempt at making the page list headings
more attractive. 

TODO: the sort action needs to be based on the current action.