use user's display_name, not login, in most page lists. (closes: #309)
I wonder why it was ever any other way, but there seems to have been a
fair amount of conscious effort to make it show only logins in a number
of cases. Perhaps this change will get reverted for we.riseup, but
UNICEF wants it and it makes enough sense to me that I'm not doing this
in the UNICEF-only branch.