no delete and move pages for non coordinators
Normal members of a group or a network that has coord access to a page
should not be able to delete or move that page.
This is a little hacky but less restrictive. If we want to go really
restrictive we might restrict all :admin actions to admins of
coordinator groups or users with direct coordinator access to the page.