making it so, that sites don't see stuff in other sites:
found out, I had to patch a lot of different actions, which list groups,
pages, assets etc.

I got a lot of them, but not every, so just find out.

used some different concepts

there is no pathfinder  that supports sites, which would be not so bad

sometimes it seems, we can get the space for the site

in Network.has?(Page) we use Page.groups_with_access?(:view) which
doesn't work, because it returns not the ones with :edit and :admin

in Network.pages_for_network we collect the complete pages for all the
groups in a network
So we can use intersection of whatever pages we have with that array to
specifiy the site range
This should be eventually replaced by a specific pathfinder

In ApplicationController we kick out everything, we should not have
access to.
That causes a problem with the pages from above

TESTCASE:

1) Every list must only contain items, that don't raise a
PermissinDeniedError
2) Every item, that
a) is direct child of the network (which are in fact none in the
testdata)
b) I have access to and is part of one group in the current_site.network
must be in the lists.

TODO:

1) rebuild the testdata with items and users in different sites, that
have nothing to do with each other
2) find things, where the filters are not applied yet
3) Get as far as possible rid of the additional queries (like
Network.pages_for_network), rewrite the pathfinder or build a cache like
that:

Network.total_page_ids