use layouts for banner and 2nd nav no more @banner
This is the big main commit of this rework. Check out the new files
in app/views/layouts - they are at the core of everything ;)

Instead of setting @banner and alike we now define layout on a per
controller / action basis. The layouts themselves are generic for
groups, users and netorks and check @active_tab to figure out what they
actually are about.