fixed posts controller. Do NOT use login twice in a single test
moved test data into fixtures instead of generating in every test

login_as does NOT remove the last user logged in - so you can not use this twice in the same test.

I tried to fix but it's more complex than I thought and you should not do this anyway. Generate your test data ahead of time and test a single action per test to isolate the tests.