getting destroy image to work.
Also removed an unclear comment above destroy action.
What I meant was we can remove the asset completely and not just from the gallery page as we now only allow images to show up on one gallery.
Before we had to take care they were not shown in another gallery.